+++ gulpfile.js
@@ -1,10 +1,15 @@
+const fs = require('fs-extra');
+const path = require('path');
+const chalk = require('chalk');
+const archiver = require('archiver');
+const stringify = require('json-stringify-pretty-compact');
 const jsdoc2md = require('jsdoc-to-markdown');
 const { parallel, series } = require('gulp');
-const fs = require('fs');
 const zip = require('gulp-zip');
 const gulp = require('gulp');
 const version = require('./package.json').version;
 const fetch = require('node-fetch');
+const argv = require('yargs').argv;
 
 function docs(done) {
   jsdoc2md
@@ -41,24 +46,36 @@ async function patrons(done) {
   return done();
 }
 
-function build(done) {
-  gulp.src('module.json').pipe(gulp.dest('dist'));
+/**
+ * Build JavaScript
+ */
+function buildJS() {
+  return (
+    gulp
+      .src([
+        'src/**/*.js',
+        '!dist/**',
+        '!out/**',
+        '!jsdoc/**',
+        '!node_modules/**',
+        '!.gitignore',
+        '!gulpfile.js',
+        '!package.json',
+        '!package-lock.json',
+      ])
 
-  gulp
-    .src([
-      '**/*',
-      '!dist/**',
-      '!out/**',
-      '!jsdoc/**',
-      '!node_modules/**',
-      '!.gitignore',
-      '!gulpfile.js',
-      '!package.json',
-      '!package-lock.json',
-    ])
-    .pipe(zip(`combat-carousel.zip`))
-    .pipe(gulp.dest('dist'));
-  return done();
+      // // eslint() attaches the lint output to the "eslint" property
+      // // of the file object so it can be used by other modules.
+      // .pipe(eslint())
+      // // eslint.format() outputs the lint results to the console.
+      // // Alternatively use eslint.formatEach() (see Docs).
+      // .pipe(eslint.format())
+      // // To have the process exit with an error code (1) on
+      // // lint error, return the stream and pipe to failAfterError last.
+      // .pipe(eslint.failAfterError())
+
+      .pipe(gulp.dest('dist'))
+  );
 }
 
 async function fetchPatrons(patrons = [], nextPage = null) {
@@ -104,10 +121,104 @@ async function fetchPatrons(patrons = [], nextPage = null) {
   return patrons;
 }
 
+/**
+ * Package build
+ */
+async function packageBuild() {
+  const manifest = getManifest();
+
+  return new Promise((resolve, reject) => {
+    try {
+      // Remove the package dir without doing anything else
+      if (argv.clean || argv.c) {
+        console.log(chalk.yellow('Removing all packaged files'));
+        fs.removeSync('package');
+        return;
+      }
+
+      // Ensure there is a directory to hold all the packaged versions
+      fs.ensureDirSync('package');
+
+      // Initialize the zip file
+      const zipName = `${manifest.file.name}-v${manifest.file.version}.zip`;
+      const zipFile = fs.createWriteStream(path.join('package', zipName));
+      const zip = archiver('zip', { zlib: { level: 9 } });
+
+      zipFile.on('close', () => {
+        console.log(chalk.green(zip.pointer() + ' total bytes'));
+        console.log(chalk.green(`Zip file ${zipName} has been written`));
+        return resolve();
+      });
+
+      zip.on('error', (err) => {
+        throw err;
+      });
+
+      zip.pipe(zipFile);
+
+      // Add the directory with the final code
+      zip.directory('dist/', manifest.file.name);
+
+      zip.finalize();
+    } catch (err) {
+      return reject(err);
+    }
+  });
+}
+
+function getManifest() {
+  const json = {};
+
+  if (fs.existsSync('src')) {
+    json.root = 'src';
+  } else {
+    json.root = 'dist';
+  }
+
+  const modulePath = path.join(json.root, 'module.json');
+  const systemPath = path.join(json.root, 'system.json');
+
+  if (fs.existsSync(modulePath)) {
+    json.file = fs.readJSONSync(modulePath);
+    json.name = 'module.json';
+  } else if (fs.existsSync(systemPath)) {
+    json.file = fs.readJSONSync(systemPath);
+    json.name = 'system.json';
+  } else {
+    return;
+  }
+
+  return json;
+}
+
+/**
+ * Build Css
+ */
+function buildCSS() {
+  return gulp.src('src/**/*.css').pipe(gulp.dest('dist'));
+}
+
+/**
+ * Copy static files
+ */
+async function copyFiles() {
+  const statics = ['lang', 'fonts', 'assets', 'icons', 'templates', 'module.json', 'system.json', 'template.json'];
+  try {
+    for (const file of statics) {
+      if (fs.existsSync(path.join('src', file))) {
+        await fs.copy(path.join('src', file), path.join('dist', file));
+      }
+    }
+    return Promise.resolve();
+  } catch (err) {
+    Promise.reject(err);
+  }
+}
+
 const chores = parallel(patrons, docs);
 
-exports.build = build;
-exports.docs = docs;
-exports.patrons = patrons;
+exports.build = gulp.series(gulp.parallel(buildJS, buildCSS, copyFiles));
+// exports.docs = docs;
+// exports.patrons = patrons;
 exports.chores = chores;
-exports.default = series(chores, build);
+exports.default = gulp.series(chores, buildJS);